# docs/swagger.yaml
basePath: /
definitions:
...
  internal_router_api_v1_auth.ChangeEmailRequest:
    properties:
      email:
        type: string
    type: object
  internal_router_api_v1_auth.ChangePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
  internal_router_api_v1_auth.ChangePhoneNumberRequest:
    properties:
      phone_number:
        type: string
    type: object
  internal_router_api_v1_auth.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
...
  internal_router_api_v1_auth.GetRefreshTokenRequest:
    properties:
      refresh_token_id:
        type: integer
    type: object
  internal_router_api_v1_auth.GetRefreshTokenResponseBody:
    properties:
      data:
        $ref: '#/definitions/internal_router_api_v1_auth.GetRefreshTokenResponseData'
      status:
        type: string
    type: object
...
  internal_router_api_v1_auth.LogInRequest:
    properties:
      is_totp_recovery_code:
        type: boolean
      password:
        type: string
      totp_code:
        type: string
      username:
        type: string
    type: object
  internal_router_api_v1_auth.ResetPasswordRequest:
    properties:
      password:
        type: string
    type: object
  internal_router_api_v1_auth.RevokeRefreshTokenRequest:
    properties:
      refresh_token_id:
        type: integer
    type: object
...
  internal_router_api_v1_auth.VerifyPhoneNumberRequest:
    properties:
      token:
        type: string
    type: object
...
  response.JSendErrorBody:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
  response.JSendFailBody:
    properties:
      code:
        type: string
      data: {}
      status:
        type: string
    type: object
  response.JSendSuccessBody:
    properties:
      data: {}
      status:
        type: string
    type: object
info:
  contact: {}
  description: This is the REST API for the Secure Notes application.
  license:
    name: GPL-3.0
    url: http://www.gnu.org/licenses/gpl-3.0.html
  title: Secure Notes REST API
  version: "1.0"
paths:
  /api/v1/auth/email:
    put:
      consumes:
      - application/json
      description: Changes the email of the authenticated user
      parameters:
      - description: Change Email Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_auth.ChangeEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Changes the email of the authenticated user
      tags:
      - api v1 auth
  /api/v1/auth/email/send-verification:
    post:
      consumes:
      - application/json
      description: Sends an email verification token to the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Sends an email verification token to the authenticated user
      tags:
      - api v1 auth
  /api/v1/auth/email/verify/{token}:
    post:
      consumes:
      - application/json
      description: Verifies the email of the authenticated user
      parameters:
      - description: Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Verifies the email of the authenticated user
      tags:
      - api v1 auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a seed token
      parameters:
      - description: Log In Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_auth.LogInRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      summary: Log in a user
      tags:
      - api v1 auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Logs out a user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Log out a user
      tags:
      - api v1 auth
  /api/v1/auth/password:
    put:
      consumes:
      - application/json
      description: Changes a user's password
      parameters:
      - description: Change Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_auth.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Change a user's password
      tags:
      - api v1 auth
  /api/v1/auth/password/forgot:
    post:
      consumes:
      - application/json
      description: Sends a password reset email
      parameters:
      - description: Forgot Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_auth.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Send a password reset email
      tags:
      - api v1 auth
  /api/v1/auth/password/reset/{token}:
    post:
      consumes:
      - application/json
      description: Resets a user's password
      parameters:
      - description: Token
        in: path
        name: token
        required: true
        type: string
      - description: Reset Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_auth.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Reset a user's password
      tags:
      - api v1 auth
  /api/v1/auth/phone-number:
    put:
      consumes:
      - application/json
      description: Changes the phone number of the authenticated user
      parameters:
      - description: Change Phone Number Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_auth.ChangePhoneNumberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Changes the phone number of the authenticated user
      tags:
      - api v1 auth
  /api/v1/auth/phone-number/send-verification:
    post:
      consumes:
      - application/json
      description: Sends a phone number verification code to the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Sends a phone number verification code to the authenticated user
      tags:
      - api v1 auth
  /api/v1/auth/phone-number/verify:
    post:
      consumes:
      - application/json
      description: Verifies the phone number of the authenticated user
      parameters:
      - description: Verify Phone Number Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_auth.VerifyPhoneNumberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Verifies the phone number of the authenticated user
      tags:
      - api v1 auth
  /api/v1/auth/refresh-token:
    delete:
      consumes:
      - application/json
      description: Revokes a user's refresh token
      parameters:
      - description: Revoke Refresh Token Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_auth.RevokeRefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Revoke a user's refresh token
      tags:
      - api v1 auth
    get:
      consumes:
      - application/json
      description: Gets a user's refresh token
      parameters:
      - description: Get Refresh Token Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_auth.GetRefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_router_api_v1_auth.GetRefreshTokenResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Get a user's refresh token
      tags:
      - api v1 auth
    post:
      consumes:
      - application/json
      description: Refreshes a user token
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Refresh a user token
      tags:
      - api v1 auth
...
  /api/v1/auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a new user account with the provided details
      parameters:
      - description: Sign Up Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_auth.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      summary: Sign up a new user
      tags:
      - api v1 auth
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"